+++ src/website/package.json
@@ -17,16 +17,9 @@
     "@dnd-kit/core": "^6.1.0",
     "@dnd-kit/modifiers": "^7.0.0",
     "@dnd-kit/sortable": "^8.0.0",
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "^11.11.0",
-    "@mui/icons-material": "^5.15.10",
-    "@mui/material": "^5.15.12",
-    "@react-google-maps/api": "^2.19.2",
     "@reduxjs/toolkit": "^1.9.5",
-    "dayjs": "^1.11.10",
     "dotenv": "^16.3.1",
     "exceljs": "^4.4.0",
-    "express": "^4.18.2",
     "leaflet": "^1.9.4",
     "leaflet-defaulticon-compatibility": "^0.1.2",
     "leaflet-geometryutil": "^0.10.2",
@@ -38,12 +31,9 @@
     "react-dom": "^18.2.0",
     "react-leaflet": "^4.2.1",
     "react-redux": "^8.1.2",
-    "react-select": "^5.8.0",
-    "recharts": "^2.10.4",
     "redux": "^4.2.1",
     "redux-logger": "^3.0.6",
     "redux-saga": "^1.2.3",
-    "redux-thunk": "^2.4.2",
     "uplot": "^1.6.30",
     "uplot-react": "^1.1.5"
   },
