--- src/website/pages/Polaris/index.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import PolarisContainer from '@/views/components/Polaris/PolarisContainer';
-
-export default function DataSets() {
-  return <PolarisContainer />;
-}
