--- src/website/pages/dashboard/index.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import DashboardContainer from '@/views/DashboardContainer';
-
-export default function Dashboard() {
-  return <DashboardContainer />;
-}
