--- src/website/views/components/BoatCompass/boatcompass.module.css
@@ -1,13 +1,0 @@
-.bottom {
-  position: absolute;
-}
-
-.middle {
-  position: absolute;
-  z-index: 1;
-}
-
-.top {
-  position: relative;
-  z-index: 1000000000;
-}
