--- src/website/views/components/Dataset/dataset.module.css
@@ -1,58 +1,0 @@
-@import 'https://fonts.googleapis.com/css2?family=Switzer:wght@400;700&display=swap';
-
-.flexItemContainer {
-  flex: 1;
-  display: flex;
-  justify-content: center;
-  font-family: Switzer, sans-serif;
-  color: #092e4e;
-  font-weight: bold;
-}
-
-.detailsFlexContainer {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-  position: relative;
-  padding-bottom: 8px;
-  margin-top: 2px;
-}
-
-.detailsFlexContainer::after {
-  content: '';
-  display: block;
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-  height: 1px;
-  background-color: black;
-}
-
-.contentContainer {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-  padding-top: 8px;
-}
-
-.accordionCustom {
-  background-color: #7a8b99;
-  border: 1px solid #bdc3c7;
-  box-shadow: none;
-}
-
-.accordionSummaryCustom {
-  background-color: #d9dddc;
-  color: #00263e;
-  font-family: Switzer, sans-serif;
-  font-weight: bold;
-}
-
-.accordionDetailsCustom {
-  background-color: white;
-  flex-direction: column;
-  padding: 16px;
-  font-family: Switzer, sans-serif;
-  color: #092e4e;
-}
