--- src/website/views/components/SingleValueChart/singlevaluechart.module.css
@@ -1,5 +1,0 @@
-.card {
-  display: block;
-  width: 15vw;
-  height: 13vh;
-}
