--- src/website/views/components/SingleValueLine/singlevalueline.module.css
@@ -1,17 +1,0 @@
-.parent {
-  position: relative;
-}
-
-.topRight {
-  position: absolute;
-  z-index: 10000000000;
-  top: 0.6vh;
-  right: 0.3vw;
-}
-
-.bottomRight {
-  position: absolute;
-  z-index: 10000000000;
-  bottom: 3.1vh;
-  right: 0.5vw;
-}
